refactor: replyComment/Comment 서비스단 분리#20
