+++ good2travel/src/main/java/com/spot/good2travel/service/CommentService.java
@@ -7,7 +7,6 @@ import com.spot.good2travel.common.exception.UserNotAuthorizedException;
 import com.spot.good2travel.common.security.CustomUserDetails;
 import com.spot.good2travel.domain.Comment;
 import com.spot.good2travel.domain.Post;
-import com.spot.good2travel.domain.ReplyComment;
 import com.spot.good2travel.domain.User;
 import com.spot.good2travel.dto.CommentResponse;
 import com.spot.good2travel.repository.CommentRepository;
@@ -24,7 +23,6 @@ import java.util.List;
 import java.util.Objects;
 
 import static com.spot.good2travel.dto.CommentRequest.CommentCreateUpdateRequest;
-import static com.spot.good2travel.dto.CommentRequest.ReplyCommentCreateUpdateRequest;
 
 
 @Service
@@ -36,7 +34,6 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final ReplyCommentRepository replyCommentRepository;
     private final RedisTemplate<String, Object> redisTemplate;
-    private final ImageService imageService;
 
 
     @Transactional
@@ -49,21 +46,6 @@ public class CommentService {
                 .findById(request.getPostId()).orElseThrow(()-> new NotFoundElementException(ExceptionMessage.POST_NOT_FOUND));
 
         commentRepository.save(Comment.of(request, user, post));
-    }
-
-    @Transactional
-    public void addReplyComment(ReplyCommentCreateUpdateRequest request, UserDetails userDetails){
-        Long userId = ((CustomUserDetails) userDetails).getId();
-
-        postRepository //게시글이 없는데 댓글을 어떻게 쓰나요!
-                .findById(request.getPostId()).orElseThrow(()-> new NotFoundElementException(ExceptionMessage.POST_NOT_FOUND));
-
-        User user = userRepository.findById(userId)
-                .orElseThrow(()-> new NotFoundElementException(ExceptionMessage.MEMBER_NOT_FOUND));
-        Comment comment = commentRepository.findById(request.getCommentId())
-                .orElseThrow(()-> new NotFoundElementException(ExceptionMessage.COMMENT_NOT_FOUND));
-
-        replyCommentRepository.save(ReplyComment.of(request, user, comment));
     }
 
     @Transactional
@@ -123,27 +105,6 @@ public class CommentService {
     }
 
     @Transactional
-    public void reportReplyComment(Long commentId, UserDetails userDetails){
-        if(userDetails == null){
-            throw new NotFoundElementException(ExceptionMessage.TOKEN_NOT_FOUND);
-        }
-
-        Long userId = ((CustomUserDetails) userDetails).getId();
-        String commentReportKey = "reportCommentId:" + commentId;
-        ReplyComment comment = replyCommentRepository.findById(commentId)
-                .orElseThrow(()-> new NotFoundElementException(ExceptionMessage.REPLY_COMMENT_NOT_FOUND));
-
-        if(Boolean.FALSE.equals(redisTemplate.opsForSet().isMember(commentReportKey, userId))){
-            redisTemplate.opsForSet().add(commentReportKey, userId);
-            Integer beforeReport = comment.getReport();
-            comment.updateReplyCommentReport(beforeReport+1);
-        }
-        else{
-            throw new UserNotAuthorizedException(ExceptionMessage.ALREADY_REPORTED);
-        }
-    }
-
-    @Transactional
     public void deleteComment(Long commentId, UserDetails userDetails){
         if(userDetails == null){
             throw new NotFoundElementException(ExceptionMessage.TOKEN_NOT_FOUND);
@@ -162,24 +123,6 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteReplyComment(Long ReplyCommentId, UserDetails userDetails){
-        if(userDetails == null){
-            throw new NotFoundElementException(ExceptionMessage.TOKEN_NOT_FOUND);
-        }
-
-        Long userId = ((CustomUserDetails) userDetails).getId();
-
-        ReplyComment comment = replyCommentRepository.findById(ReplyCommentId)
-                .orElseThrow(()->new NotFoundElementException(ExceptionMessage.REPLY_COMMENT_NOT_FOUND));
-
-        if(!Objects.equals(comment.getUser().getId(), userId)){
-            throw new UserNotAuthorizedException(ExceptionMessage.COMMENT_NOT_USERS);
-        }
-
-        replyCommentRepository.delete(comment);
-    }
-
-    @Transactional
     public void updateComment(Long commentId, CommentCreateUpdateRequest request, UserDetails userDetails){
         if(userDetails == null){
             throw new NotFoundElementException(ExceptionMessage.TOKEN_NOT_FOUND);
@@ -192,22 +135,6 @@ public class CommentService {
 
         if(comment.getUser().getId().equals(userId)){
             comment.updateComment(request);
-        }
-    }
-
-    @Transactional
-    public void updateReplyComment(Long replyCommentId, ReplyCommentCreateUpdateRequest request, UserDetails userDetails){
-        if(userDetails == null){
-            throw new NotFoundElementException(ExceptionMessage.TOKEN_NOT_FOUND);
-        }
-
-        Long userId = ((CustomUserDetails) userDetails).getId();
-
-        ReplyComment comment = replyCommentRepository.findById(replyCommentId)
-                .orElseThrow(()->new NotFoundElementException(ExceptionMessage.REPLY_COMMENT_NOT_FOUND));
-
-        if(comment.getUser().getId().equals(userId)){
-            comment.updateReplyComment(request);
         }
     }
